refactor(arch): decompose giant screens and unify stats engine delegates
