+++ app/src/main/java/com/vinnovateit/latch/features/settings/SettingsScreen.kt
@@ -44,7 +44,6 @@ import androidx.compose.material.icons.rounded.SettingsBackupRestore
 import androidx.compose.material.icons.rounded.SettingsSystemDaydream
 import androidx.compose.material.icons.rounded.Speed
 import androidx.compose.material.icons.rounded.Vibration
-import androidx.compose.material3.AlertDialog
 import androidx.compose.material3.Button
 import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.FilledIconButton
@@ -53,8 +52,6 @@ import androidx.compose.material3.IconButtonDefaults
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.ModalBottomSheet
 import androidx.compose.material3.OutlinedButton
-import androidx.compose.material3.OutlinedTextField
-import androidx.compose.material3.Slider
 import androidx.compose.material3.Surface
 import androidx.compose.material3.Switch
 import androidx.compose.material3.Text
@@ -93,6 +90,8 @@ import com.vinnovateit.latch.R
 import com.vinnovateit.latch.common.ui.components.ExpressiveTopBarContent
 import com.vinnovateit.latch.platform.LatchAppGraph
 import com.vinnovateit.latch.features.settings.manager.SettingsManager
+import com.vinnovateit.latch.features.settings.components.CustomColorPickerDialog
+import com.vinnovateit.latch.features.settings.components.parseHexOrNull
 import kotlinx.coroutines.launch
 import kotlin.math.roundToInt
 
@@ -707,56 +706,6 @@ fun SettingsSection(
   }
 }
 
-private const val CustomColorSaturation = 0.75f
-private const val CustomColorLightness = 0.45f
-
-private fun hslToColor(hue: Float, saturation: Float = CustomColorSaturation, lightness: Float = CustomColorLightness): Color {
-  val c = (1f - kotlin.math.abs(2f * lightness - 1f)) * saturation
-  val x = c * (1f - kotlin.math.abs((hue / 60f) % 2f - 1f))
-  val m = lightness - c / 2f
-  val (r1, g1, b1) = when {
-    hue < 60f -> Triple(c, x, 0f)
-    hue < 120f -> Triple(x, c, 0f)
-    hue < 180f -> Triple(0f, c, x)
-    hue < 240f -> Triple(0f, x, c)
-    hue < 300f -> Triple(x, 0f, c)
-    else -> Triple(c, 0f, x)
-  }
-  return Color(r1 + m, g1 + m, b1 + m)
-}
-
-private fun colorToHue(color: Color): Float {
-  val r = color.red
-  val g = color.green
-  val b = color.blue
-  val max = maxOf(r, g, b)
-  val min = minOf(r, g, b)
-  val delta = max - min
-  if (delta == 0f) return 0f
-  val hue = when (max) {
-    r -> 60f * (((g - b) / delta) % 6f)
-    g -> 60f * (((b - r) / delta) + 2f)
-    else -> 60f * (((r - g) / delta) + 4f)
-  }
-  return if (hue < 0f) hue + 360f else hue
-}
-
-internal fun parseHexOrNull(value: String): Color? {
-  return try {
-    val hex = value.removePrefix("#")
-    if (hex.length == 6) Color(0xFF000000.toInt() or hex.toInt(16)) else null
-  } catch (_: Exception) {
-    null
-  }
-}
-
-private fun Color.toHexString(): String {
-  val r = (red * 255).roundToInt().coerceIn(0, 255)
-  val g = (green * 255).roundToInt().coerceIn(0, 255)
-  val b = (blue * 255).roundToInt().coerceIn(0, 255)
-  return String.format("#%02X%02X%02X", r, g, b)
-}
-
 @Composable
 fun AccentColorPicker(
   selectedColorName: String,
@@ -846,181 +795,6 @@ fun AccentColorPicker(
       }
     )
   }
-}
-
-@Composable
-fun CustomColorPickerDialog(
-  initialColorHex: String,
-  onDismiss: () -> Unit,
-  onColorConfirmed: (String) -> Unit
-) {
-  var hexInput by remember { mutableStateOf(initialColorHex.uppercase()) }
-  val initialColor = parseHexOrNull(initialColorHex) ?: Color(0xFFC01221)
-  var currentColor by remember { mutableStateOf(initialColor) }
-  var hue by remember { mutableFloatStateOf(colorToHue(initialColor)) }
-
-  val formattedHex = if (hexInput.startsWith("#")) hexInput else "#$hexInput"
-  val isValidHex = parseHexOrNull(formattedHex) != null
-
-  AlertDialog(
-    onDismissRequest = onDismiss,
-    title = {
-      Text(
-        text = "Custom Accent Color",
-        style = MaterialTheme.typography.titleLarge,
-        fontWeight = FontWeight.Bold
-      )
-    },
-    text = {
-      Column(
-        modifier = Modifier
-          .fillMaxWidth()
-          .verticalScroll(rememberScrollState()),
-        horizontalAlignment = Alignment.CenterHorizontally
-      ) {
-        // Preview Swatch
-        Box(
-          modifier = Modifier
-            .fillMaxWidth()
-            .height(56.dp)
-            .clip(RoundedCornerShape(12.dp))
-            .background(currentColor)
-            .border(1.dp, MaterialTheme.colorScheme.outlineVariant, RoundedCornerShape(12.dp)),
-          contentAlignment = Alignment.Center
-        ) {
-          val isLight = (0.299 * currentColor.red + 0.587 * currentColor.green + 0.114 * currentColor.blue) > 0.5
-          Text(
-            text = currentColor.toHexString(),
-            color = if (isLight) Color.Black else Color.White,
-            fontWeight = FontWeight.Bold,
-            style = MaterialTheme.typography.titleMedium
-          )
-        }
-
-        Spacer(modifier = Modifier.height(20.dp))
-
-        // Hue Slider
-        Text(
-          text = "Hue",
-          style = MaterialTheme.typography.labelMedium,
-          fontWeight = FontWeight.SemiBold,
-          color = MaterialTheme.colorScheme.onSurfaceVariant,
-          modifier = Modifier.fillMaxWidth()
-        )
-
-        Spacer(modifier = Modifier.height(6.dp))
-
-        val rainbowBrush = remember {
-          Brush.horizontalGradient(
-            (0..360 step 30).map { hslToColor(it.toFloat()) }
-          )
-        }
-
-        Box(
-          modifier = Modifier
-            .fillMaxWidth()
-            .height(14.dp)
-            .clip(RoundedCornerShape(7.dp))
-            .background(rainbowBrush)
-        )
-
-        Slider(
-          value = hue,
-          onValueChange = { newHue ->
-            hue = newHue
-            val newCol = hslToColor(newHue)
-            currentColor = newCol
-            hexInput = newCol.toHexString()
-          },
-          valueRange = 0f..360f,
-          modifier = Modifier.fillMaxWidth()
-        )
-
-        Spacer(modifier = Modifier.height(12.dp))
-
-        // Hex input
-        OutlinedTextField(
-          value = hexInput,
-          onValueChange = { input ->
-            val filtered = input.filter { it.isDigit() || it in 'a'..'f' || it in 'A'..'F' || it == '#' }.take(7)
-            hexInput = filtered.uppercase()
-            val candidate = if (filtered.startsWith("#")) filtered else "#$filtered"
-            parseHexOrNull(candidate)?.let { parsed ->
-              currentColor = parsed
-              hue = colorToHue(parsed)
-            }
-          },
-          label = { Text("Hex Code") },
-          placeholder = { Text("#FF5722") },
-          singleLine = true,
-          isError = hexInput.isNotEmpty() && !isValidHex,
-          supportingText = if (hexInput.isNotEmpty() && !isValidHex) {
-            { Text("Enter valid 6-digit hex (e.g. #FF5722)") }
-          } else null,
-          modifier = Modifier.fillMaxWidth(),
-          shape = RoundedCornerShape(12.dp)
-        )
-
-        Spacer(modifier = Modifier.height(16.dp))
-
-        // Preset Swatches
-        Text(
-          text = "Quick Presets",
-          style = MaterialTheme.typography.labelMedium,
-          fontWeight = FontWeight.SemiBold,
-          color = MaterialTheme.colorScheme.onSurfaceVariant,
-          modifier = Modifier.fillMaxWidth()
-        )
-
-        Spacer(modifier = Modifier.height(8.dp))
-
-        val quickPresets = listOf(
-          Color(0xFFFF5722), // Deep Orange
-          Color(0xFF009688), // Teal
-          Color(0xFF3F51B5), // Indigo
-          Color(0xFF00BCD4), // Cyan
-          Color(0xFFFF9800), // Orange
-          Color(0xFFE91E63)  // Pink
-        )
-
-        Row(
-          modifier = Modifier.fillMaxWidth(),
-          horizontalArrangement = Arrangement.SpaceBetween
-        ) {
-          quickPresets.forEach { presetColor ->
-            Box(
-              modifier = Modifier
-                .size(36.dp)
-                .clip(CircleShape)
-                .background(presetColor)
-                .clickable {
-                  currentColor = presetColor
-                  hexInput = presetColor.toHexString()
-                  hue = colorToHue(presetColor)
-                }
-            )
-          }
-        }
-      }
-    },
-    confirmButton = {
-      Button(
-        onClick = {
-          if (isValidHex) {
-            onColorConfirmed(formattedHex)
-          }
-        },
-        enabled = isValidHex
-      ) {
-        Text("Apply")
-      }
-    },
-    dismissButton = {
-      OutlinedButton(onClick = onDismiss) {
-        Text("Cancel")
-      }
-    }
-  )
 }
 
 @Composable
