+++ desktop/src/commonMain/kotlin/com/vinnovateit/latch/ui/screens/stats/components/DesktopLiveSessionCard.kt
@@ -1,0 +1,118 @@
+package com.vinnovateit.latch.ui.screens.stats.components
+
+import androidx.compose.foundation.BorderStroke
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.layout.width
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material3.Card
+import androidx.compose.material3.CardDefaults
+import androidx.compose.material3.Icon
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.LaunchedEffect
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableLongStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.setValue
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.graphics.vector.ImageVector
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.unit.dp
+import androidx.lifecycle.compose.collectAsStateWithLifecycle
+import com.vinnovateit.latch.core.model.DataUsage
+import com.vinnovateit.latch.core.settings.SettingsManager
+import com.vinnovateit.latch.core.stats.formatBitsPerSecond
+import com.vinnovateit.latch.core.stats.formatDurationDynamic
+import com.vinnovateit.latch.ui.components.DataUsageDonut
+import com.vinnovateit.latch.ui.components.LatchIcons
+import com.vinnovateit.latch.ui.theme.LocalIsDarkTheme
+import kotlinx.coroutines.delay
+
+@Composable
+fun LiveSessionCard(
+    startTimeMillis: Long,
+    usage: DataUsage,
+    latestRxBps: Long,
+    latestTxBps: Long,
+    speedUnit: String,
+    dlColor: Color,
+    ulColor: Color,
+) {
+    val usePureBlack by SettingsManager.usePureBlack.collectAsStateWithLifecycle()
+    val isAmoled = usePureBlack && LocalIsDarkTheme.current
+
+    var duration by remember(startTimeMillis) {
+        mutableLongStateOf(System.currentTimeMillis() - startTimeMillis)
+    }
+    LaunchedEffect(startTimeMillis) {
+        while (true) {
+            duration = System.currentTimeMillis() - startTimeMillis
+            delay(1000)
+        }
+    }
+
+    Card(
+        modifier = Modifier.fillMaxWidth().padding(horizontal = 16.dp),
+        shape = RoundedCornerShape(28.dp),
+        colors = CardDefaults.cardColors(
+            containerColor = MaterialTheme.colorScheme.surfaceContainerHighest,
+        ),
+        elevation = CardDefaults.cardElevation(defaultElevation = 8.dp),
+        border = if (isAmoled) BorderStroke(4.dp, MaterialTheme.colorScheme.primary) else null,
+    ) {
+        Column(modifier = Modifier.fillMaxWidth().padding(20.dp)) {
+            Row(
+                modifier = Modifier.fillMaxWidth(),
+                horizontalArrangement = Arrangement.SpaceBetween,
+                verticalAlignment = Alignment.CenterVertically,
+            ) {
+                Text(
+                    text = "Active Session",
+                    style = MaterialTheme.typography.titleMedium,
+                    fontWeight = FontWeight.Bold,
+                    color = MaterialTheme.colorScheme.onSurface,
+                )
+                Text(
+                    text = formatDurationDynamic(duration),
+                    style = MaterialTheme.typography.bodySmall,
+                    color = MaterialTheme.colorScheme.onSurfaceVariant,
+                )
+            }
+            Spacer(Modifier.height(16.dp))
+            Row(modifier = Modifier.fillMaxWidth(), verticalAlignment = Alignment.CenterVertically) {
+                DataUsageDonut(data = usage, modifier = Modifier.size(96.dp), isAmoled = isAmoled, dlColor = dlColor, ulColor = ulColor)
+                Spacer(Modifier.width(24.dp))
+                Column(modifier = Modifier.weight(1f), verticalArrangement = Arrangement.spacedBy(10.dp)) {
+                    val (rxValue, rxUnit) = formatBitsPerSecond(latestRxBps, speedUnit)
+                    val (txValue, txUnit) = formatBitsPerSecond(latestTxBps, speedUnit)
+                    RateChip(LatchIcons.ArrowUpward, "$txValue $txUnit", ulColor)
+                    RateChip(LatchIcons.ArrowDownward, "$rxValue $rxUnit", dlColor)
+                }
+            }
+        }
+    }
+}
+
+@Composable
+private fun RateChip(icon: ImageVector, text: String, accent: Color) {
+    Row(verticalAlignment = Alignment.CenterVertically) {
+        Icon(icon, contentDescription = null, tint = accent, modifier = Modifier.size(16.dp))
+        Spacer(Modifier.width(4.dp))
+        Text(
+            text = text,
+            style = MaterialTheme.typography.labelLarge,
+            fontWeight = FontWeight.Bold,
+            color = MaterialTheme.colorScheme.onSurface,
+        )
+    }
+}
